requirements.txt
Flask==2.1.0
Pillow>=8.0.0
torch
easyocr

